Split up old View component into Nav and Notice
