+++ src/NoticeClass.tsx
